requirements.txt
streamlit
tuya-connector-python
requests